backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
